task_4/requirements.txt
# Kogniti Minds CRM Lite - Requirements
# Install with: pip install -r requirements.txt

pandas>=1.3.0
matplotlib>=3.5.0
openpyxl>=3.0.0
fpdf2>=2.5.0
sqlite3

# Optional for enhanced features
plotly>=5.0.0  # For interactive charts
seaborn>=0.11.0  # For better visualizations